requirements.txt
streamlit>=1.31.0
plotly>=5.18.0
google-generativeai>=0.3.2
python-dotenv>=1.0.0
requests>=2.31.0
groq>=0.4.2
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0